// src/types/album.d.ts
export declare type AlbumInfo = {
  name: string;
  alias: string[];
  briefDesc: string;
  mark: number;
  id: number;
  description: string;
  picUrl: string;
  subType: string;
  songs: any[];
  transNames: string[];
  type: string;
  company: string;
  onSale: boolean;
  publishTime: number;
  tags: string;

  //额外加的
  publishTimeStr: string;
};